refactor(Request): 인터페이스 제거, Request 단일 객체 활용
