+++ src/main/java/chess/view/InputView.java
@@ -4,8 +4,6 @@ import static camp.nextstep.edu.missionutils.Console.*;
 
 import java.util.List;
 
-import chess.dto.MoveRequest;
-import chess.dto.NotMoveRequest;
 import chess.dto.Request;
 import chess.model.Position;
 import chess.vo.Command;
@@ -23,11 +21,11 @@ public class InputView {
         String input = readLine();
         Command command = Command.MoveStatusEnd(input);
         if (command == Command.END || command == Command.STATUS) {
-            return new NotMoveRequest(command);
+            return new Request(command);
         }
 
         List<String> inputs = List.of(input.split(" "));
-        return new MoveRequest(command, Position.of(inputs.get(1)), Position.of(inputs.get(2)));
+        return new Request(command, Position.of(inputs.get(1)), Position.of(inputs.get(2)));
     }
 
 }
